requirements.txt
flask
python-gitlab